Navigation to the new skills page works as expected.
It was observed that each time a server component was rendered, a request was made to the API irregardless of any previous requests made to the endpoint. This is due to the fact that requests from prefetchQuery(), do not hit the client side cache, as this method is only meant to be used on the server.

So, a single instance cache will be implemented to prevent unnecessary requests from server components to the API.