The category options in the dropdown are now provided by data from the database.
There was an issue where the requests for these categories was returning null; this was due to server-side authorization which were a result of an invalid access token being attached to each request. This problem arose because we were trying to use the client query function on the server (this is an issue because the server doesn't have access to browser cookies - the place where the access token is stored).