Skill cards have been completed.
The next steps it to implement the modal that will be used to modify the skill cards (use similar to approach to AddSkillModal - maybe generalize the component).

Once the above task is completed, research how to use ReactQuery with Mutations for optimistic updates (to ensure smoother experience for users - a possible approach involves making a request after each route change with the updated data . This would be strictly for user-specific data like profile and skills).